added new index.js for with upload component
