+++ src/components/pspdfkit/index-original.js
@@ -1,0 +1,61 @@
+import React, { Component } from 'react';
+import PSPDFKitWeb from 'pspdfkit';
+
+export default class PSPDFKit extends Component {
+	constructor(props, context) {
+		super(props, context);
+		this._instance = null;
+		this._container = null;
+
+		this.onRef = this.onRef.bind(this);
+		this.load = this.load.bind(this);
+		this.unload = this.unload.bind(this);
+	}
+
+	onRef(container) {
+		this._container = container;
+	}
+
+	async load(props) {
+		console.log(`Loading ${props.documentUrl}`);
+
+		this._instance = await PSPDFKitWeb.load({
+			document: props.documentUrl,
+			container: this._container,
+			baseUrl: props.baseUrl,
+		});
+		console.log('Successfully mounted PSPDFKit', this._instance);
+	}
+
+	unload() {
+		PSPDFKitWeb.unload(this._instance || this._container);
+		this._instance = null;
+	}
+
+	componentDidMount() {
+		this.load(this.props);
+	}
+
+	componentDidUpdate(prevProps) {
+		const nextProps = this.props;
+
+		// We only want to reload the document when the documentUrl prop changes.
+		if (nextProps.documentUrl !== prevProps.documentUrl) {
+			this.unload();
+			this.load(nextProps);
+		}
+	}
+
+	componentWillUnmount() {
+		this.unload();
+	}
+
+	render() {
+		return (
+			<div
+				ref={this.onRef}
+				style={{ width: '100%', height: '100%', position: 'absolute' }}
+			/>
+		);
+	}
+}
